Put out initial project methods
Enclosed are my project methods for the basic fetch and arithmetic operations involved in the web site. More can be added later. The methods themselves are implemented as middleware for now, but each one has a corresponding "kernel" function which actually does the work. Since those functions are not being served as middleware, they are free to be used as building blocks for more advanced methods and, as you will see on review of the methods, they are already being used as well. I have also included some basic test js files I used to perform unit testing. testapp.js serves up the middleware on a machine with Port 3456 set up, while testdata.js is simply a few collections for test data. Those will eventually be replaced by MongoDB queries.