Add additional APIs for testing
Of particular interest should be the API for '/allUserPortInfo'. We can change that name to whatever we want, but if we're trying to cut back on the API endpoints, then that one is our big workhorse that will give the front end all the info it needs for a user's portfolio in one go.