Updated project-methods for better versions
I have rewritten the project-methods.js (and testapp.js) to be more compliant with the final version. It's still serving up data off of testdata.js, which must be replaced with calls to MongoDB, but now the APIs (unless we want to change the exact URLs) should be able to hook up with the front end.